+++ lib/coinbase/client/models/transaction.rb
@@ -180,7 +180,7 @@ module Coinbase::Client
       return false if @from_address_id.nil?
       return false if @unsigned_payload.nil?
       return false if @status.nil?
-      status_validator = EnumAttributeValidator.new('String', ["pending", "broadcast", "complete", "failed"])
+      status_validator = EnumAttributeValidator.new('String', ["pending", "signed", "broadcast", "complete", "failed"])
       return false unless status_validator.valid?(@status)
       true
     end
@@ -188,7 +188,7 @@ module Coinbase::Client
     # Custom attribute writer method checking allowed values (enum).
     # @param [Object] status Object to be assigned
     def status=(status)
-      validator = EnumAttributeValidator.new('String', ["pending", "broadcast", "complete", "failed"])
+      validator = EnumAttributeValidator.new('String', ["pending", "signed", "broadcast", "complete", "failed"])
       unless validator.valid?(status)
         fail ArgumentError, "invalid value for \"status\", must be one of #{validator.allowable_values}."
       end