+++ lib/coinbase/client/models/validator.rb
@@ -14,22 +14,30 @@ require 'date'
 require 'time'
 
 module Coinbase::Client
-  class CreateServerSignerRequest
-    # The ID of the server signer for the 1 of 1 server signer.
-    attr_accessor :server_signer_id
+  # A validator onchain.
+  class Validator
+    # The publicly identifiable unique id of the validator. This can be the public key for Ethereum validators and maybe an address for some other network.
+    attr_accessor :validator_id
 
-    # The enrollment data of the server signer. This will be the base64 encoded server-signer-id for the 1 of 1 server signer.
-    attr_accessor :enrollment_data
+    # The ID of the blockchain network to which the Validator belongs.
+    attr_accessor :network_id
 
-    # Whether the Server-Signer uses MPC.
-    attr_accessor :is_mpc
+    # The ID of the asset that the validator helps stake.
+    attr_accessor :asset_id
+
+    # The status of the validator.
+    attr_accessor :status
+
+    attr_accessor :details
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'server_signer_id' => :'server_signer_id',
-        :'enrollment_data' => :'enrollment_data',
-        :'is_mpc' => :'is_mpc'
+        :'validator_id' => :'validator_id',
+        :'network_id' => :'network_id',
+        :'asset_id' => :'asset_id',
+        :'status' => :'status',
+        :'details' => :'details'
       }
     end
 
@@ -41,9 +49,11 @@ module Coinbase::Client
     # Attribute type mapping.
     def self.openapi_types
       {
-        :'server_signer_id' => :'String',
-        :'enrollment_data' => :'String',
-        :'is_mpc' => :'Boolean'
+        :'validator_id' => :'String',
+        :'network_id' => :'String',
+        :'asset_id' => :'String',
+        :'status' => :'String',
+        :'details' => :'ValidatorDetails'
       }
     end
 
@@ -57,31 +67,43 @@ module Coinbase::Client
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::CreateServerSignerRequest` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Coinbase::Client::Validator` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::CreateServerSignerRequest`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Coinbase::Client::Validator`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'server_signer_id')
-        self.server_signer_id = attributes[:'server_signer_id']
+      if attributes.key?(:'validator_id')
+        self.validator_id = attributes[:'validator_id']
+      else
+        self.validator_id = nil
       end
 
-      if attributes.key?(:'enrollment_data')
-        self.enrollment_data = attributes[:'enrollment_data']
+      if attributes.key?(:'network_id')
+        self.network_id = attributes[:'network_id']
       else
-        self.enrollment_data = nil
+        self.network_id = nil
       end
 
-      if attributes.key?(:'is_mpc')
-        self.is_mpc = attributes[:'is_mpc']
+      if attributes.key?(:'asset_id')
+        self.asset_id = attributes[:'asset_id']
       else
-        self.is_mpc = nil
+        self.asset_id = nil
+      end
+
+      if attributes.key?(:'status')
+        self.status = attributes[:'status']
+      else
+        self.status = nil
+      end
+
+      if attributes.key?(:'details')
+        self.details = attributes[:'details']
       end
     end
 
@@ -90,12 +112,20 @@ module Coinbase::Client
     def list_invalid_properties
       warn '[DEPRECATED] the `list_invalid_properties` method is obsolete'
       invalid_properties = Array.new
-      if @enrollment_data.nil?
-        invalid_properties.push('invalid value for "enrollment_data", enrollment_data cannot be nil.')
+      if @validator_id.nil?
+        invalid_properties.push('invalid value for "validator_id", validator_id cannot be nil.')
       end
 
-      if @is_mpc.nil?
-        invalid_properties.push('invalid value for "is_mpc", is_mpc cannot be nil.')
+      if @network_id.nil?
+        invalid_properties.push('invalid value for "network_id", network_id cannot be nil.')
+      end
+
+      if @asset_id.nil?
+        invalid_properties.push('invalid value for "asset_id", asset_id cannot be nil.')
+      end
+
+      if @status.nil?
+        invalid_properties.push('invalid value for "status", status cannot be nil.')
       end
 
       invalid_properties
@@ -105,8 +135,10 @@ module Coinbase::Client
     # @return true if the model is valid
     def valid?
       warn '[DEPRECATED] the `valid?` method is obsolete'
-      return false if @enrollment_data.nil?
-      return false if @is_mpc.nil?
+      return false if @validator_id.nil?
+      return false if @network_id.nil?
+      return false if @asset_id.nil?
+      return false if @status.nil?
       true
     end
 
@@ -115,9 +147,11 @@ module Coinbase::Client
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          server_signer_id == o.server_signer_id &&
-          enrollment_data == o.enrollment_data &&
-          is_mpc == o.is_mpc
+          validator_id == o.validator_id &&
+          network_id == o.network_id &&
+          asset_id == o.asset_id &&
+          status == o.status &&
+          details == o.details
     end
 
     # @see the `==` method
@@ -129,7 +163,7 @@ module Coinbase::Client
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [server_signer_id, enrollment_data, is_mpc].hash
+      [validator_id, network_id, asset_id, status, details].hash
     end
 
     # Builds the object from hash