+++ lib/coinbase/staking_reward.rb
@@ -47,10 +47,16 @@ module Coinbase
       @model.date
     end
 
+    # Returns the onchain address of the StakingReward.
+    # @return [Time] The onchain address
+    def address_id
+      @model.address_id
+    end
+
     # Returns a string representation of the StakingReward.
     # @return [String] a string representation of the StakingReward
     def to_s
-      "Coinbase::StakingReward{amount: '#{amount}'}"
+      "Coinbase::StakingReward{date: '#{date}' address_id: '#{address_id}' amount: '#{amount.to_f}'}"
     end
 
     # Same as to_s.